requirements.txt
google-cloud-pubsub
google-cloud-bigquery
google-cloud-logging